[Fix] #41 - AnnotationView 따로 파일로 분리해서정리
